+++ CHANGELOG.md
@@ -46,7 +46,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * [#8169](https://github.com/osmosis-labs/osmosis/pull/8169) Support non-pool assets in superfluid staking
 * [#8274](https://github.com/osmosis-labs/osmosis/pull/8274) SDK v50 and Comet v0.38 upgrade
-* [#8310](https://github.com/osmosis-labs/osmosis/pull/8310) Taker fee share agreement logic added
+* [#8310](https://github.com/osmosis-labs/osmosis/pull/8310) Taker fee share
 
 ### State Compatible